Rename oso visibility permission to "query"
Oso gives special treatment to one permission. Normally, if the agent
does not have the required permission, oso.authorize() raises a
ForbiddenError, but if the agent doesn't even have this special
permission, then oso.authorize() raises a NotFoundError and the
application should pretend the resource doesn't exist. By default, this
special permission is named "read". To better correspond with the
GraphQL API, this commit changes the special permission to "query".

Signed-off-by: John Arnold <code@iohannes.us>