Refactor GQL loading functions
Previous to this commit, SQLAlchemy's native eager loading functionality
was used to eagerly load a parent class's child relationships while
walking the GQL query. This took place in the function
prepare_orm_query(). Now the various ORM objects are loaded by standard
select statements in the function recursive_eager_load().

This change is to allow a foundation for future implementation of
filtering and limiting return per-parent, which SQLAlchemy's native
eager loading functionality is not designed to do.

Signed-off-by: John Arnold <code@iohannes.us>