Add SQL CTE in eager loader to limit # of results
When loading GraphQL queries of the form:
`parents { children(first: 10) { id } }`
it is not possible to use SQL LIMIT, because we do not want 10 children
in total but rather 10 children per parent. To do that, it is necessary
to use an SQL window function in a CTE, which this commit implements.

Signed-off-by: John Arnold <code@iohannes.us>