Move to new Associate model
This commit implements a new database and GraphQL model, Associate.
This replaces the ProspectiveResident model, and changes the role of the
User model. Now, the Associate model represents any person a community
has a connection to; i.e. people as objects, while the User model
represents a user account; i.e. people as subjects.

Signed-off-by: John Arnold <code@iohannes.us>